+++ player/tv/src/main/kotlin/dev/aaa1115910/bv/player/tv/controller/ControllerVideoInfo.kt
@@ -250,7 +250,7 @@ fun ControllerVideoInfoBottom(
     stateData: VideoPlayerStateData,
     idleIcon: String,
     movingIcon: String,
-    play: Int,
+    play: Long,
     danmaku: Int,
     like: Int,
     coin: Int,
